Famme sti test

+++ bookkeeper-server/src/test/java/org/apache/bookkeeper/bookie/storage/ldb/WriteCacheTest.java
@@ -87,7 +87,7 @@ public class WriteCacheTest {
         }
         for(int c = 0; c <10; c ++) {
         	long d = (long)0 + (long)c;
-             assertEquals(entry, cache.get(1.0,1.0));
+             assertEquals(entry, cache.get((long)1,(long)1));
              assertEquals((cache.get(d,d)),entry);
              cache.deleteLedger(d);
              assertEquals(cache.get(d,d), null);
